// frontend/src/external/client/api/orval/generated/zod/scout-templates/scout-templates.zod.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * inselfy API
 * OpenAPI spec version: 1.0.0
 */
import * as zod from 'zod';

import { zodParams } from '../../../zod-params';

/**
 * スカウトテンプレート作成（企業認証）
 * @summary Create a scout template
 */
export const scoutTemplatesCreateScoutTemplateBodyNameMax = 100;

export const scoutTemplatesCreateScoutTemplateBodySubjectMax = 200;

export const scoutTemplatesCreateScoutTemplateBodyBodyMax = 5000;



export const ScoutTemplatesCreateScoutTemplateBody = zod.object({
  "name": zod.string(zodParams({"operationId":"ScoutTemplates_createScoutTemplate","location":"body","schemaName":"ScoutTemplatesCreateScoutTemplateBody","fieldPath":["name"],"validator":"string"})).max(scoutTemplatesCreateScoutTemplateBodyNameMax, zodParams({"operationId":"ScoutTemplates_createScoutTemplate","location":"body","schemaName":"ScoutTemplatesCreateScoutTemplateBody","fieldPath":["name"],"validator":"max"})).describe('テンプレート名'),
  "subject": zod.string(zodParams({"operationId":"ScoutTemplates_createScoutTemplate","location":"body","schemaName":"ScoutTemplatesCreateScoutTemplateBody","fieldPath":["subject"],"validator":"string"})).max(scoutTemplatesCreateScoutTemplateBodySubjectMax, zodParams({"operationId":"ScoutTemplates_createScoutTemplate","location":"body","schemaName":"ScoutTemplatesCreateScoutTemplateBody","fieldPath":["subject"],"validator":"max"})).describe('件名'),
  "body": zod.string(zodParams({"operationId":"ScoutTemplates_createScoutTemplate","location":"body","schemaName":"ScoutTemplatesCreateScoutTemplateBody","fieldPath":["body"],"validator":"string"})).max(scoutTemplatesCreateScoutTemplateBodyBodyMax, zodParams({"operationId":"ScoutTemplates_createScoutTemplate","location":"body","schemaName":"ScoutTemplatesCreateScoutTemplateBody","fieldPath":["body"],"validator":"max"})).describe('本文')
}).describe('スカウトテンプレート作成リクエスト')

/**
 * スカウトテンプレート更新（企業認証）
 * @summary Update a scout template
 */
export const scoutTemplatesUpdateScoutTemplateBodyNameMax = 100;

export const scoutTemplatesUpdateScoutTemplateBodySubjectMax = 200;

export const scoutTemplatesUpdateScoutTemplateBodyBodyMax = 5000;



export const ScoutTemplatesUpdateScoutTemplateBody = zod.object({
  "name": zod.string(zodParams({"operationId":"ScoutTemplates_updateScoutTemplate","location":"body","schemaName":"ScoutTemplatesUpdateScoutTemplateBody","fieldPath":["name"],"validator":"string"})).max(scoutTemplatesUpdateScoutTemplateBodyNameMax, zodParams({"operationId":"ScoutTemplates_updateScoutTemplate","location":"body","schemaName":"ScoutTemplatesUpdateScoutTemplateBody","fieldPath":["name"],"validator":"max"})).describe('テンプレート名'),
  "subject": zod.string(zodParams({"operationId":"ScoutTemplates_updateScoutTemplate","location":"body","schemaName":"ScoutTemplatesUpdateScoutTemplateBody","fieldPath":["subject"],"validator":"string"})).max(scoutTemplatesUpdateScoutTemplateBodySubjectMax, zodParams({"operationId":"ScoutTemplates_updateScoutTemplate","location":"body","schemaName":"ScoutTemplatesUpdateScoutTemplateBody","fieldPath":["subject"],"validator":"max"})).describe('件名'),
  "body": zod.string(zodParams({"operationId":"ScoutTemplates_updateScoutTemplate","location":"body","schemaName":"ScoutTemplatesUpdateScoutTemplateBody","fieldPath":["body"],"validator":"string"})).max(scoutTemplatesUpdateScoutTemplateBodyBodyMax, zodParams({"operationId":"ScoutTemplates_updateScoutTemplate","location":"body","schemaName":"ScoutTemplatesUpdateScoutTemplateBody","fieldPath":["body"],"validator":"max"})).describe('本文')
}).describe('スカウトテンプレート更新リクエスト')